+++ src/fields/BaseRelationField.php
@@ -225,19 +225,8 @@ abstract class BaseRelationField extends Field implements PreviewableFieldInterf
         }
 
         // Config normalization
-        $nullables = [
-            'maxRelations',
-            'minRelations',
-            'selectionLabel',
-            'selectionLabel',
-            'source',
-            'targetSiteId',
-            'viewMode',
-        ];
-        foreach ($nullables as $name) {
-            if (($config[$name] ?? null) === '') {
-                unset($config[$name]);
-            }
+        if (($config['source'] ?? null) === '') {
+            unset($config['source']);
         }
 
         if (array_key_exists('sources', $config) && empty($config['sources'])) {