+++ src/mail/transportadapters/Sendmail.php
@@ -34,31 +34,18 @@ class Sendmail extends BaseTransportAdapter
 
     /**
      * @inheritdoc
-     */
-    public function __construct($config = [])
-    {
-        // Config normalization
-        if (($config['command'] ?? null) === '') {
-            unset($config['command']);
-        }
-
-        parent::__construct($config);
-    }
-
-    /**
-     * @inheritdoc
      * @since 3.4.0
      */
-    public function behaviors(): array
+    protected function defineBehaviors(): array
     {
-        $behaviors = parent::behaviors();
-        $behaviors['parser'] = [
-            'class' => EnvAttributeParserBehavior::class,
-            'attributes' => [
-                'command',
+        return [
+            'parser' => [
+                'class' => EnvAttributeParserBehavior::class,
+                'attributes' => [
+                    'command',
+                ],
             ],
         ];
-        return $behaviors;
     }
 
     /**